operations/ipquery/requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.2.1
Flask-Login==0.2.11
Flask-WTF==0.11
Flask-Cache==0.13.1
pysaml2==2.4.0
pyyaml==3.11
ip2instance>=1.2.0
